Update: Added image and fixed the responsiveness
## What?
I added an image and fixed the responsiveness of this project.

## Why?
This is a practice I made to close the CSS module in my META course.

## How?
In this update I tried to practice grids and flexbox. Also I put on practice the max and min width properties. Also I tried to use a minimalist design.

## Testing?

## Screenshots

## Anything Else? (optional)
I had troubles with the footer and trying to put a display fixed but the content of the main element was hidden by it.

I learned how to make responsible images as the one is included on the page.